Extract panel header to title bar (#625)
